Isolated more tests to linux only
