+++ kotest-framework/kotest-framework-engine/src/jvmTest/kotlin/com/sksamuel/kotest/engine/extensions/LazyMountableExtensionTest.kt
@@ -1,5 +1,7 @@
 package com.sksamuel.kotest.engine.extensions
 
+import io.kotest.core.annotation.EnabledIf
+import io.kotest.core.annotation.enabledif.LinuxCondition
 import io.kotest.core.extensions.LazyMaterialized
 import io.kotest.core.extensions.LazyMountableExtension
 import io.kotest.core.extensions.install
@@ -7,6 +9,7 @@ import io.kotest.core.spec.style.FunSpec
 import io.kotest.matchers.shouldBe
 import kotlinx.coroutines.delay
 
+@EnabledIf(LinuxCondition::class)
 class LazyMountableExtensionTest : FunSpec() {
 
    private val mountable = MyLazyMountableExtension()
@@ -19,6 +22,7 @@ class LazyMountableExtensionTest : FunSpec() {
    }
 }
 
+@EnabledIf(LinuxCondition::class)
 class MyLazyMountableExtension : LazyMountableExtension<Unit, String> {
 
    override fun mount(configure: (Unit) -> Unit): LazyMaterialized<String> {
