+++ kotest-framework/kotest-framework-engine/src/jvmTest/kotlin/com/sksamuel/kotest/engine/extensions/spec/finalizespec/FinalizeSpecListenerTest.kt
@@ -1,5 +1,7 @@
 package com.sksamuel.kotest.engine.extensions.spec.finalizespec
 
+import io.kotest.core.annotation.EnabledIf
+import io.kotest.core.annotation.enabledif.LinuxCondition
 import io.kotest.core.config.ProjectConfiguration
 import io.kotest.core.listeners.FinalizeSpecListener
 import io.kotest.core.listeners.TestListener
@@ -31,6 +33,7 @@ private class FinalizeSpecTestListener2 : FinalizeSpecListener {
    }
 }
 
+@EnabledIf(LinuxCondition::class)
 class FinalizeSpecTest : FunSpec() {
    init {
       test("finalize spec listeners should be fired") {
@@ -50,6 +53,7 @@ class FinalizeSpecTest : FunSpec() {
    }
 }
 
+@EnabledIf(LinuxCondition::class)
 class FinalizeSpec : FunSpec() {
    init {
       test("ignored test").config(enabled = false) {}
