+++ kotest-framework/kotest-framework-engine/src/jvmTest/kotlin/com/sksamuel/kotest/engine/extensions/spec/specextensions/LateinitSpecExtensionStringSpecTest.kt
@@ -1,10 +1,13 @@
 package com.sksamuel.kotest.engine.extensions.spec.specextensions
 
+import io.kotest.core.annotation.EnabledIf
+import io.kotest.core.annotation.enabledif.LinuxCondition
 import io.kotest.core.extensions.SpecExtension
 import io.kotest.core.spec.Spec
 import io.kotest.core.spec.style.StringSpec
 import io.kotest.matchers.shouldBe
 
+@EnabledIf(LinuxCondition::class)
 class LateinitSpecExtensionStringSpecTest : StringSpec() {
 
    private lateinit var string: String
