+++ kotest-framework/kotest-framework-engine/src/jvmTest/kotlin/com/sksamuel/kotest/engine/extensions/spec/specextensions/LateinitSpecExtensionWordSpecTest.kt
@@ -1,10 +1,13 @@
 package com.sksamuel.kotest.engine.extensions.spec.specextensions
 
+import io.kotest.core.annotation.EnabledIf
+import io.kotest.core.annotation.enabledif.LinuxCondition
 import io.kotest.core.extensions.SpecExtension
 import io.kotest.core.spec.Spec
 import io.kotest.core.spec.style.WordSpec
 import io.kotest.matchers.shouldBe
 
+@EnabledIf(LinuxCondition::class)
 class LateinitSpecExtensionWordSpecTest : WordSpec() {
 
    private lateinit var string: String
