+++ kotest-framework/kotest-framework-engine/src/jvmTest/kotlin/com/sksamuel/kotest/engine/extensions/test/AfterAnyExceptionTest.kt
@@ -1,6 +1,8 @@
 package com.sksamuel.kotest.engine.extensions.test
 
 import io.kotest.core.Platform
+import io.kotest.core.annotation.EnabledIf
+import io.kotest.core.annotation.enabledif.LinuxCondition
 import io.kotest.core.config.ProjectConfiguration
 import io.kotest.core.spec.IsolationMode
 import io.kotest.core.spec.style.BehaviorSpec
@@ -48,7 +50,7 @@ private class StringSpecWithAfterTestError : StringSpec({
    afterTest {
       error("boom")
    }
-   "string test"{}
+   "string test" {}
 })
 
 private class ShouldSpecWithAfterTestError : ShouldSpec({
@@ -101,6 +103,7 @@ private class WordSpecWithAfterTestError : WordSpec({
    }
 })
 
+@EnabledIf(LinuxCondition::class)
 class AfterAnyExceptionTest : WordSpec({
 
    var error: Throwable? = null
@@ -114,55 +117,82 @@ class AfterAnyExceptionTest : WordSpec({
 
    "an exception in before test" should {
       "fail the test for behavior spec" {
-         val executor = SpecExecutor(NoopCoroutineDispatcherFactory, EngineContext(ProjectConfiguration(), Platform.JVM).withListener(listener))
+         val executor = SpecExecutor(
+            NoopCoroutineDispatcherFactory,
+            EngineContext(ProjectConfiguration(), Platform.JVM).withListener(listener)
+         )
          executor.execute(BehaviorSpecWithAfterTestError::class)
          error.shouldBeInstanceOf<ExtensionException.AfterAnyException>()
          error!!.cause!!.shouldHaveMessage("boom")
       }
       "fail the test for feature spec" {
-         val executor = SpecExecutor(NoopCoroutineDispatcherFactory, EngineContext(ProjectConfiguration(), Platform.JVM).withListener(listener))
+         val executor = SpecExecutor(
+            NoopCoroutineDispatcherFactory,
+            EngineContext(ProjectConfiguration(), Platform.JVM).withListener(listener)
+         )
          executor.execute(FeatureSpecWithAfterTestError::class)
          error.shouldBeInstanceOf<ExtensionException.AfterAnyException>()
          error!!.cause!!.shouldHaveMessage("boom")
       }
       "fail the test for word spec" {
-         val executor = SpecExecutor(NoopCoroutineDispatcherFactory, EngineContext(ProjectConfiguration(), Platform.JVM).withListener(listener))
+         val executor = SpecExecutor(
+            NoopCoroutineDispatcherFactory,
+            EngineContext(ProjectConfiguration(), Platform.JVM).withListener(listener)
+         )
          executor.execute(WordSpecWithAfterTestError::class)
          error.shouldBeInstanceOf<ExtensionException.AfterAnyException>()
          error!!.cause!!.shouldHaveMessage("boom")
       }
       "fail the test for should spec" {
-         val executor = SpecExecutor(NoopCoroutineDispatcherFactory, EngineContext(ProjectConfiguration(), Platform.JVM).withListener(listener))
+         val executor = SpecExecutor(
+            NoopCoroutineDispatcherFactory,
+            EngineContext(ProjectConfiguration(), Platform.JVM).withListener(listener)
+         )
          executor.execute(ShouldSpecWithAfterTestError::class)
          error.shouldBeInstanceOf<ExtensionException.AfterAnyException>()
          error!!.cause!!.shouldHaveMessage("boom")
       }
       "fail the test for string spec" {
-         val executor = SpecExecutor(NoopCoroutineDispatcherFactory, EngineContext(ProjectConfiguration(), Platform.JVM).withListener(listener))
+         val executor = SpecExecutor(
+            NoopCoroutineDispatcherFactory,
+            EngineContext(ProjectConfiguration(), Platform.JVM).withListener(listener)
+         )
          executor.execute(StringSpecWithAfterTestError::class)
          error.shouldBeInstanceOf<ExtensionException.AfterAnyException>()
          error!!.cause!!.shouldHaveMessage("boom")
       }
       "fail the test for describe spec" {
-         val executor = SpecExecutor(NoopCoroutineDispatcherFactory, EngineContext(ProjectConfiguration(), Platform.JVM).withListener(listener))
+         val executor = SpecExecutor(
+            NoopCoroutineDispatcherFactory,
+            EngineContext(ProjectConfiguration(), Platform.JVM).withListener(listener)
+         )
          executor.execute(DescribeSpecWithAfterTestError::class)
          error.shouldBeInstanceOf<ExtensionException.AfterAnyException>()
          error!!.cause!!.shouldHaveMessage("boom")
       }
       "fail the test for free spec" {
-         val executor = SpecExecutor(NoopCoroutineDispatcherFactory, EngineContext(ProjectConfiguration(), Platform.JVM).withListener(listener))
+         val executor = SpecExecutor(
+            NoopCoroutineDispatcherFactory,
+            EngineContext(ProjectConfiguration(), Platform.JVM).withListener(listener)
+         )
          executor.execute(FreeSpecWithAfterTestError::class)
          error.shouldBeInstanceOf<ExtensionException.AfterAnyException>()
          error!!.cause!!.shouldHaveMessage("boom")
       }
       "fail the test for fun spec" {
-         val executor = SpecExecutor(NoopCoroutineDispatcherFactory, EngineContext(ProjectConfiguration(), Platform.JVM).withListener(listener))
+         val executor = SpecExecutor(
+            NoopCoroutineDispatcherFactory,
+            EngineContext(ProjectConfiguration(), Platform.JVM).withListener(listener)
+         )
          executor.execute(FunSpecWithAfterTestError::class)
          error.shouldBeInstanceOf<ExtensionException.AfterAnyException>()
          error!!.cause!!.shouldHaveMessage("boom")
       }
       "fail the test for expect spec" {
-         val executor = SpecExecutor(NoopCoroutineDispatcherFactory, EngineContext(ProjectConfiguration(), Platform.JVM).withListener(listener))
+         val executor = SpecExecutor(
+            NoopCoroutineDispatcherFactory,
+            EngineContext(ProjectConfiguration(), Platform.JVM).withListener(listener)
+         )
          executor.execute(ExpectSpecWithAfterTestError::class)
          error.shouldBeInstanceOf<ExtensionException.AfterAnyException>()
          error!!.cause!!.shouldHaveMessage("boom")
