+++ kotest-framework/kotest-framework-engine/src/jvmTest/kotlin/com/sksamuel/kotest/engine/extensions/test/AfterContainerDescribeSpecTest.kt
@@ -1,32 +1,11 @@
 package com.sksamuel.kotest.engine.extensions.test
 
+import io.kotest.core.annotation.EnabledIf
+import io.kotest.core.annotation.enabledif.LinuxCondition
 import io.kotest.core.spec.style.DescribeSpec
-import io.kotest.core.spec.style.FunSpec
 import io.kotest.matchers.shouldBe
 
-class AfterContainerFunSpecTest : FunSpec() {
-   var a = ""
-
-   init {
-
-      afterContainer {
-         a += "2"
-      }
-
-      context("container1") {
-         test("a") {}
-      }
-
-      context("container2") {
-         test("a") {}
-      }
-
-      afterProject {
-         a shouldBe "22"
-      }
-   }
-}
-
+@EnabledIf(LinuxCondition::class)
 class AfterContainerDescribeSpecTest : DescribeSpec() {
    var a = ""
