+++ kotest-framework/kotest-framework-engine/src/jvmTest/kotlin/com/sksamuel/kotest/engine/extensions/test/BeforeAnyExceptionTest.kt
@@ -1,6 +1,8 @@
 package com.sksamuel.kotest.engine.extensions.test
 
 import io.kotest.core.Platform
+import io.kotest.core.annotation.EnabledIf
+import io.kotest.core.annotation.enabledif.LinuxCondition
 import io.kotest.core.config.ProjectConfiguration
 import io.kotest.core.spec.IsolationMode
 import io.kotest.core.spec.style.BehaviorSpec
@@ -51,7 +53,7 @@ private class StringSpecWithBeforeTestError : StringSpec({
    beforeTest {
       error("boom")
    }
-   "string test"{}
+   "string test" {}
 })
 
 private class ShouldSpecWithBeforeTestError : ShouldSpec({
@@ -104,6 +106,7 @@ private class WordSpecWithBeforeTestError : WordSpec({
    }
 })
 
+@EnabledIf(LinuxCondition::class)
 class BeforeAnyExceptionTest : WordSpec({
 
    var error: Throwable? = null
@@ -117,55 +120,82 @@ class BeforeAnyExceptionTest : WordSpec({
 
    "an exception in before test" should {
       "fail the test for behavior spec" {
-         val executor = SpecExecutor(NoopCoroutineDispatcherFactory, EngineContext(ProjectConfiguration(), Platform.JVM).withListener(listener))
+         val executor = SpecExecutor(
+            NoopCoroutineDispatcherFactory,
+            EngineContext(ProjectConfiguration(), Platform.JVM).withListener(listener)
+         )
          executor.execute(BehaviorSpecWithBeforeTestError::class)
          error.shouldBeInstanceOf<ExtensionException.BeforeAnyException>()
          error!!.cause!!.shouldHaveMessage("boom")
       }
       "fail the test for feature spec" {
-         val executor = SpecExecutor(NoopCoroutineDispatcherFactory, EngineContext(ProjectConfiguration(), Platform.JVM).withListener(listener))
+         val executor = SpecExecutor(
+            NoopCoroutineDispatcherFactory,
+            EngineContext(ProjectConfiguration(), Platform.JVM).withListener(listener)
+         )
          executor.execute(FeatureSpecWithBeforeTestError::class)
          error.shouldBeInstanceOf<ExtensionException.BeforeAnyException>()
          error!!.cause!!.shouldHaveMessage("boom")
       }
       "fail the test for word spec" {
-         val executor = SpecExecutor(NoopCoroutineDispatcherFactory, EngineContext(ProjectConfiguration(), Platform.JVM).withListener(listener))
+         val executor = SpecExecutor(
+            NoopCoroutineDispatcherFactory,
+            EngineContext(ProjectConfiguration(), Platform.JVM).withListener(listener)
+         )
          executor.execute(WordSpecWithBeforeTestError::class)
          error.shouldBeInstanceOf<ExtensionException.BeforeAnyException>()
          error!!.cause!!.shouldHaveMessage("boom")
       }
       "fail the test for should spec" {
-         val executor = SpecExecutor(NoopCoroutineDispatcherFactory, EngineContext(ProjectConfiguration(), Platform.JVM).withListener(listener))
+         val executor = SpecExecutor(
+            NoopCoroutineDispatcherFactory,
+            EngineContext(ProjectConfiguration(), Platform.JVM).withListener(listener)
+         )
          executor.execute(ShouldSpecWithBeforeTestError::class)
          error.shouldBeInstanceOf<ExtensionException.BeforeAnyException>()
          error!!.cause!!.shouldHaveMessage("boom")
       }
       "fail the test for string spec" {
-         val executor = SpecExecutor(NoopCoroutineDispatcherFactory, EngineContext(ProjectConfiguration(), Platform.JVM).withListener(listener))
+         val executor = SpecExecutor(
+            NoopCoroutineDispatcherFactory,
+            EngineContext(ProjectConfiguration(), Platform.JVM).withListener(listener)
+         )
          executor.execute(StringSpecWithBeforeTestError::class)
          error.shouldBeInstanceOf<ExtensionException.BeforeAnyException>()
          error!!.cause!!.shouldHaveMessage("boom")
       }
       "fail the test for describe spec" {
-         val executor = SpecExecutor(NoopCoroutineDispatcherFactory, EngineContext(ProjectConfiguration(), Platform.JVM).withListener(listener))
+         val executor = SpecExecutor(
+            NoopCoroutineDispatcherFactory,
+            EngineContext(ProjectConfiguration(), Platform.JVM).withListener(listener)
+         )
          executor.execute(DescribeSpecWithBeforeTestError::class)
          error.shouldBeInstanceOf<ExtensionException.BeforeAnyException>()
          error!!.cause!!.shouldHaveMessage("boom")
       }
       "fail the test for free spec" {
-         val executor = SpecExecutor(NoopCoroutineDispatcherFactory, EngineContext(ProjectConfiguration(), Platform.JVM).withListener(listener))
+         val executor = SpecExecutor(
+            NoopCoroutineDispatcherFactory,
+            EngineContext(ProjectConfiguration(), Platform.JVM).withListener(listener)
+         )
          executor.execute(FreeSpecWithBeforeTestError::class)
          error.shouldBeInstanceOf<ExtensionException.BeforeAnyException>()
          error!!.cause!!.shouldHaveMessage("boom")
       }
       "fail the test for fun spec" {
-         val executor = SpecExecutor(NoopCoroutineDispatcherFactory, EngineContext(ProjectConfiguration(), Platform.JVM).withListener(listener))
+         val executor = SpecExecutor(
+            NoopCoroutineDispatcherFactory,
+            EngineContext(ProjectConfiguration(), Platform.JVM).withListener(listener)
+         )
          executor.execute(FunSpecWithBeforeTestError::class)
          error.shouldBeInstanceOf<ExtensionException.BeforeAnyException>()
          error!!.cause!!.shouldHaveMessage("boom")
       }
       "fail the test for expect spec" {
-         val executor = SpecExecutor(NoopCoroutineDispatcherFactory, EngineContext(ProjectConfiguration(), Platform.JVM).withListener(listener))
+         val executor = SpecExecutor(
+            NoopCoroutineDispatcherFactory,
+            EngineContext(ProjectConfiguration(), Platform.JVM).withListener(listener)
+         )
          executor.execute(ExpectSpecWithBeforeTestError::class)
          error.shouldBeInstanceOf<ExtensionException.BeforeAnyException>()
          error!!.cause!!.shouldHaveMessage("boom")
