+++ kotest-framework/kotest-framework-engine/src/jvmTest/kotlin/com/sksamuel/kotest/engine/interceptors/DumpConfigInterceptorTest.kt
@@ -1,5 +1,7 @@
 package com.sksamuel.kotest.engine.interceptors
 
+import io.kotest.core.annotation.EnabledIf
+import io.kotest.core.annotation.enabledif.LinuxCondition
 import io.kotest.core.config.ProjectConfiguration
 import io.kotest.core.spec.style.FunSpec
 import io.kotest.datatest.withData
@@ -10,6 +12,7 @@ import io.kotest.extensions.system.SystemOutWireListener
 import io.kotest.matchers.shouldBe
 import io.kotest.matchers.string.shouldBeEmpty
 
+@EnabledIf(LinuxCondition::class)
 class DumpConfigInterceptorTest : FunSpec({
 
    val property = "kotest.framework.dump.config"
