+++ kotest-framework/kotest-framework-engine/src/jvmTest/kotlin/com/sksamuel/kotest/engine/interceptors/ProjectExtensionEngineInterceptorTest.kt
@@ -1,17 +1,18 @@
 package com.sksamuel.kotest.engine.interceptors
 
-import io.kotest.common.KotestInternal
 import io.kotest.core.TagExpression
+import io.kotest.core.annotation.EnabledIf
+import io.kotest.core.annotation.enabledif.LinuxCondition
 import io.kotest.core.config.ProjectConfiguration
-import io.kotest.core.project.ProjectContext
 import io.kotest.core.extensions.ProjectExtension
+import io.kotest.core.project.ProjectContext
 import io.kotest.core.spec.style.FunSpec
 import io.kotest.engine.EngineResult
 import io.kotest.engine.interceptors.EngineContext
 import io.kotest.engine.interceptors.ProjectExtensionEngineInterceptor
 import io.kotest.matchers.shouldBe
 
-@KotestInternal
+@EnabledIf(LinuxCondition::class)
 class ProjectExtensionEngineInterceptorTest : FunSpec({
 
    test("should invoke all extensions") {
