+++ kotest-framework/kotest-framework-engine/src/jvmTest/kotlin/com/sksamuel/kotest/engine/interceptors/SpecFilterInterceptorTest.kt
@@ -1,5 +1,7 @@
 package com.sksamuel.kotest.engine.interceptors
 
+import io.kotest.core.annotation.EnabledIf
+import io.kotest.core.annotation.enabledif.LinuxCondition
 import io.kotest.core.config.ProjectConfiguration
 import io.kotest.core.filter.SpecFilter
 import io.kotest.core.filter.SpecFilterResult
@@ -11,6 +13,7 @@ import io.kotest.matchers.booleans.shouldBeFalse
 import io.kotest.matchers.booleans.shouldBeTrue
 import kotlin.reflect.KClass
 
+@EnabledIf(LinuxCondition::class)
 class SpecFilterInterceptorTest : FunSpec() {
    init {
       test("spec filter's should filter tests") {
