+++ kotest-framework/kotest-framework-engine/src/jvmTest/kotlin/com/sksamuel/kotest/engine/interceptors/TestEngineStartedFinishedInterceptorTest.kt
@@ -1,6 +1,7 @@
 package com.sksamuel.kotest.engine.interceptors
 
-import io.kotest.common.KotestInternal
+import io.kotest.core.annotation.EnabledIf
+import io.kotest.core.annotation.enabledif.LinuxCondition
 import io.kotest.core.spec.style.FunSpec
 import io.kotest.engine.EngineResult
 import io.kotest.engine.interceptors.EngineContext
@@ -9,7 +10,7 @@ import io.kotest.engine.listener.AbstractTestEngineListener
 import io.kotest.matchers.collections.shouldHaveSize
 import io.kotest.matchers.shouldBe
 
-@KotestInternal
+@EnabledIf(LinuxCondition::class)
 class TestEngineStartedFinishedInterceptorTest : FunSpec({
 
    test("should invoke engineStarted before downstream") {
