+++ kotest-framework/kotest-framework-engine/src/jvmTest/kotlin/com/sksamuel/kotest/engine/spec/annotation/AnnotationSpecBeforeAfterTest.kt
@@ -1,5 +1,7 @@
 package com.sksamuel.kotest.engine.spec.annotation
 
+import io.kotest.core.annotation.EnabledIf
+import io.kotest.core.annotation.enabledif.LinuxCondition
 import io.kotest.core.extensions.ApplyExtension
 import io.kotest.core.listeners.TestListener
 import io.kotest.core.spec.IsolationMode
@@ -14,6 +16,7 @@ import java.util.concurrent.atomic.AtomicInteger
 import kotlin.reflect.KClass
 
 @ApplyExtension(AssertionListener::class)
+@EnabledIf(LinuxCondition::class)
 class AnnotationSpecBeforeAfterTest : AnnotationSpec() {
 
    companion object {
