+++ kotest-framework/kotest-framework-engine/src/jvmTest/kotlin/com/sksamuel/kotest/engine/spec/annotation/AnnotationSpecIgnoresTest.kt
@@ -1,12 +1,15 @@
 package com.sksamuel.kotest.engine.spec.annotation
 
 import io.kotest.assertions.fail
+import io.kotest.core.annotation.EnabledIf
+import io.kotest.core.annotation.enabledif.LinuxCondition
 import io.kotest.core.spec.style.AnnotationSpec
 import io.kotest.core.spec.style.DescribeSpec
 import io.kotest.engine.TestEngineLauncher
 import io.kotest.engine.listener.CollectingTestEngineListener
 import io.kotest.matchers.shouldBe
 
+@EnabledIf(LinuxCondition::class)
 class AnnotationSpecIgnoresTest : DescribeSpec({
 
    describe("An AnnotationSpec") {
