+++ kotest-framework/kotest-framework-engine/src/jvmTest/kotlin/com/sksamuel/kotest/engine/spec/config/DescribeSpecConfigSyntaxTest.kt
@@ -2,6 +2,8 @@ package com.sksamuel.kotest.engine.spec.config
 
 import io.kotest.common.ExperimentalKotest
 import io.kotest.common.nonConstantFalse
+import io.kotest.core.annotation.EnabledIf
+import io.kotest.core.annotation.enabledif.LinuxCondition
 import io.kotest.core.spec.style.DescribeSpec
 import io.kotest.core.test.Enabled
 import io.kotest.core.test.config.TestConfig
@@ -13,7 +15,7 @@ import kotlin.time.Duration.Companion.seconds
  * A test that just ensures the syntax for test configs does not break between releases.
  * The actual functionality of things like tags and timeouts is tested elsewhere.
  */
-@ExperimentalKotest
+@EnabledIf(LinuxCondition::class)
 class DescribeSpecConfigSyntaxTest : DescribeSpec() {
    init {
