+++ kotest-framework/kotest-framework-engine/src/jvmTest/kotlin/com/sksamuel/kotest/engine/spec/config/FunSpecConfigSyntaxTest.kt
@@ -1,8 +1,9 @@
 package com.sksamuel.kotest.engine.spec.config
 
-import io.kotest.common.ExperimentalKotest
 import io.kotest.common.nonConstantFalse
 import io.kotest.core.Tag
+import io.kotest.core.annotation.EnabledIf
+import io.kotest.core.annotation.enabledif.LinuxCondition
 import io.kotest.core.spec.style.FunSpec
 import io.kotest.core.test.config.TestConfig
 import io.kotest.matchers.shouldBe
@@ -16,7 +17,7 @@ object Tag2 : Tag()
  * A test that just ensures the syntax for test configs does not break between releases.
  * The actual functionality of things like tags and timeouts is tested elsewhere.
  */
-@ExperimentalKotest
+@EnabledIf(LinuxCondition::class)
 class FunSpecConfigSyntaxTest : FunSpec() {
    init {
