+++ kotest-framework/kotest-framework-engine/src/jvmTest/kotlin/com/sksamuel/kotest/engine/spec/config/ShouldSpecConfigSyntaxTest.kt
@@ -1,7 +1,8 @@
 package com.sksamuel.kotest.engine.spec.config
 
-import io.kotest.common.ExperimentalKotest
 import io.kotest.common.nonConstantFalse
+import io.kotest.core.annotation.EnabledIf
+import io.kotest.core.annotation.enabledif.LinuxCondition
 import io.kotest.core.spec.style.ShouldSpec
 import io.kotest.matchers.shouldBe
 import java.util.concurrent.atomic.AtomicInteger
@@ -11,7 +12,7 @@ import kotlin.time.Duration.Companion.seconds
  * A test that just ensures the syntax for test configs does not break between releases.
  * The actual functionality of things like tags and timeouts is tested elsewhere.
  */
-@ExperimentalKotest
+@EnabledIf(LinuxCondition::class)
 class ShouldSpecConfigSyntaxTest : ShouldSpec() {
    init {
