+++ kotest-framework/kotest-framework-engine/src/jvmTest/kotlin/com/sksamuel/kotest/engine/spec/coroutine/DescribeSpecCoroutineTest.kt
@@ -1,9 +1,12 @@
 package com.sksamuel.kotest.engine.spec.coroutine
 
+import io.kotest.core.annotation.EnabledIf
+import io.kotest.core.annotation.enabledif.LinuxCondition
 import io.kotest.core.spec.style.DescribeSpec
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.launch
 
+@EnabledIf(LinuxCondition::class)
 class DescribeSpecCoroutineTest : DescribeSpec() {
 
    init {
