+++ kotest-framework/kotest-framework-engine/src/jvmTest/kotlin/com/sksamuel/kotest/engine/tags/TagFilteredDiscoveryExtensionExampleTest.kt
@@ -2,7 +2,9 @@ package com.sksamuel.kotest.engine.tags
 
 import io.kotest.assertions.fail
 import io.kotest.core.NamedTag
+import io.kotest.core.annotation.EnabledIf
 import io.kotest.core.annotation.Tags
+import io.kotest.core.annotation.enabledif.LinuxCondition
 import io.kotest.core.config.ProjectConfiguration
 import io.kotest.core.extensions.TagExtension
 import io.kotest.core.spec.style.StringSpec
@@ -10,12 +12,10 @@ import io.kotest.engine.TestEngineLauncher
 import io.kotest.engine.listener.CollectingTestEngineListener
 import io.kotest.matchers.booleans.shouldBeFalse
 
+@EnabledIf(LinuxCondition::class)
 class TagFilteredDiscoveryExtensionExampleTest : StringSpec() {
    companion object {
-      val ext = object : TagExtension {
-         override fun tags(): io.kotest.core.TagExpression =
-            io.kotest.core.TagExpression(emptySet(), setOf(NamedTag("SpecExcluded")))
-      }
+      val ext = TagExtension { io.kotest.core.TagExpression(emptySet(), setOf(NamedTag("SpecExcluded"))) }
    }
 
    init {
